TIL-TodayILearned/jpashopPractice1/src/main/java/jpabook/jpashopPractice1/service/OrderService.java
package jpabook.jpashopPractice1.service;

import jpabook.jpashopPractice1.domain.*;
import jpabook.jpashopPractice1.domain.item.Item;
import jpabook.jpashopPractice1.repository.ItemRepository;
import jpabook.jpashopPractice1.repository.MemberRepository;
import jpabook.jpashopPractice1.repository.OrderRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

import java.util.List;

@Service
@Transactional(readOnly = true)
@RequiredArgsConstructor
public class OrderService {

    private final MemberRepository memberRepository;
    private final OrderRepository orderRepository;
    private final ItemRepository itemRepository;

    @Transactional
    public Long order(Long memberId, Long itemId, int count) {

        //lookup entity
        Member member = memberRepository.findOne(memberId);
        Item item = itemRepository.findOne(itemId);

        //create delivery information
        Delivery delivery = new Delivery();
        delivery.setAddress(member.getAddress());
        delivery.setStatus(DeliveryStatus.READY);

        //create order item
        OrderItem orderItem = OrderItem.createOrderItem(item, item.getPrice(), count);

        //create order
        Order order = Order.createOrder(member, delivery, orderItem);

        //store order
        orderRepository.save(order);
        return order.getId();
    }

    @Transactional
    public void cancelOrder(Long orderId) {

        //lookup order entity
        Order order = orderRepository.findOne(orderId);

        //cancel order
        order.cancel();
    }

    public List<Order> findOrders(OrderSearch orderSearch) {
        return orderRepository.findAllByString(orderSearch);
    }
}
